finalize the plugin, set to v1_0, set the API
