+++ plugins/basicmessage_storage/basicmessage_storage/definition.py
@@ -2,15 +2,9 @@
 
 versions = [
     {
-        "major_version": 0,
-        "minimum_minor_version": 1,
-        "current_minor_version": 1,
-        "path": "v0_1",
-    },
-    {
-        "major_version": 0,
-        "minimum_minor_version": 2,
-        "current_minor_version": 1,
-        "path": "v0_2",
+        "major_version": 1,
+        "minimum_minor_version": 0,
+        "current_minor_version": 0,
+        "path": "v1_0",
     }
 ]
