--- plugins/basicmessage_storage/basicmessage_storage/v0_2/__init__.py
@@ -1,35 +1,0 @@
-import logging
-import re
-
-from aries_cloudagent.config.injection_context import InjectionContext
-from aries_cloudagent.core.event_bus import EventBus, Event
-from aries_cloudagent.core.profile import Profile
-from aries_cloudagent.core.protocol_registry import ProtocolRegistry
-
-from ..v0_1.models import BasicMessageRecord
-
-LOGGER = logging.getLogger(__name__)
-
-BASIC_MESSAGE_EVENT_PATTERN = re.compile("^acapy::basicmessage::received$")
-
-
-async def setup(context: InjectionContext):
-    LOGGER.info("> plugin setup...")
-    protocol_registry = context.inject(ProtocolRegistry)
-    assert protocol_registry
-
-    event_bus = context.inject(EventBus)
-    if not event_bus:
-        raise ValueError("EventBus missing in context")
-
-    event_bus.subscribe(BASIC_MESSAGE_EVENT_PATTERN, basic_message_event_handler)
-    LOGGER.info("< plugin setup.")
-
-
-async def basic_message_event_handler(profile: Profile, event: Event):
-    # LOGGER.info(event.payload)
-    # grab the received event and persist it.
-    msg: BasicMessageRecord = BasicMessageRecord.deserialize(event.payload)
-    msg.state = BasicMessageRecord.STATE_RECV
-    async with profile.session() as session:
-        await msg.save(session, reason="New received message")
