+++ plugins/basicmessage_storage/basicmessage_storage/v1_0/__init__.py
@@ -5,8 +5,6 @@ from aries_cloudagent.config.injection_context import InjectionContext
 from aries_cloudagent.core.event_bus import EventBus, Event
 from aries_cloudagent.core.profile import Profile
 from aries_cloudagent.core.protocol_registry import ProtocolRegistry
-from aries_cloudagent.protocols.basicmessage.definition import versions
-from aries_cloudagent.protocols.basicmessage.v1_0.message_types import MESSAGE_TYPES
 
 from .models import BasicMessageRecord
 
@@ -20,20 +18,6 @@ async def setup(context: InjectionContext):
     protocol_registry = context.inject(ProtocolRegistry)
     assert protocol_registry
 
-    # this does not work...
-    # plugin registry ends up firing 'RuntimeError: OrderedDict mutated during iteration'
-    # we will have to block the basicmessage plugin in configuration
-
-    # unregister basicmessage, manually register message types not the routes
-    # plugin_registry = context.inject(PluginRegistry)
-    # assert plugin_registry
-    # plugin_registry._plugins.pop("aries_cloudagent.protocols.basicmessage")
-
-    # register basicmessage message types...
-    protocol_registry.register_message_types(
-        MESSAGE_TYPES, version_definition=versions[0]
-    )
-
     event_bus = context.inject(EventBus)
     if not event_bus:
         raise ValueError("EventBus missing in context")
@@ -43,9 +27,11 @@ async def setup(context: InjectionContext):
 
 
 async def basic_message_event_handler(profile: Profile, event: Event):
-    # LOGGER.info(event.payload)
+    LOGGER.info(event.payload)
     # grab the received event and persist it.
     msg: BasicMessageRecord = BasicMessageRecord.deserialize(event.payload)
     msg.state = BasicMessageRecord.STATE_RECV
+    LOGGER.info(msg)
     async with profile.session() as session:
         await msg.save(session, reason="New received message")
+        LOGGER.info(msg)
